+++ lib/correios/cep/address.ex
@@ -3,7 +3,7 @@ defmodule Correios.CEP.Address do
   Address structure.
   """
 
-  @enforce_keys [:street, :complement, :neighborhood, :city, :state, :postal_code, :zipcode]
+  @enforce_keys [:street, :complement, :neighborhood, :city, :state, :postal_code]
   defstruct @enforce_keys
 
   @type t() :: %__MODULE__{
@@ -12,8 +12,7 @@ defmodule Correios.CEP.Address do
           neighborhood: String.t(),
           city: String.t(),
           state: String.t(),
-          postal_code: String.t(),
-          zipcode: String.t()
+          postal_code: String.t()
         }
   @typep complement :: charlist() | nil
 
@@ -40,8 +39,7 @@ defmodule Correios.CEP.Address do
         neighborhood: "Neighborhood",
         state: "ST",
         street: "Street",
-        postal_code: "12345678",
-        zipcode: "12345678"
+        postal_code: "12345678"
       }
 
   """
@@ -61,7 +59,6 @@ defmodule Correios.CEP.Address do
       city: to_string(city),
       state: to_string(state),
       postal_code: to_string(postal_code),
-      zipcode: to_string(postal_code),
       complement: build_complement(complement1, complement2)
     }
   end