+++ test/correios/cep/integration_test.exs
@@ -13,8 +13,7 @@ defmodule Correios.CEP.IntegrationTest do
       neighborhood: "Bela Vista",
       city: "São Paulo",
       state: "SP",
-      postal_code: "01311200",
-      zipcode: "01311200"
+      postal_code: "01311200"
     }
 
     not_found_error = %Error{